requirements.txt
# Core ML
torch>=2.0
torchaudio>=2.0
numpy>=1.24

# Transcription
openai-whisper>=20231117

# Speaker diarization
pyannote.audio>=3.1

# Semantic segmentation
sentence-transformers>=2.2
transformers>=4.36
accelerate>=0.25
bitsandbytes>=0.41
scipy>=1.10

# Audio I/O
soundfile>=0.12

# Alignment / punctuation
deepmultilingualpunctuation>=1.0

# Output formatting
srt>=3.5
